% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Country Codes and Names}
\format{
A \code{data.frame} with 244 rows and 2 variables:
\describe{
\item{map_name}{Country names as defined in \code{faunabr::world_fauna}.}
\item{country_code}{Country codes used in the Catálogo Taxonômico da Fauna
do Brasil.}
}
}
\usage{
data(country_codes)
}
\description{
A dataset containing country codes used in the Catálogo Taxonômico da Fauna
do Brasil along with their corresponding country names, as defined in
\code{faunabr::world_fauna}.
}
\keyword{datasets}
