% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_translation}
\alias{map_translation}
\title{Helpers for translating data}
\format{
A \code{list} with 5 data.frames ("lifeForm", "origin", "habitat",
"taxonRank", and "taxonomicStatus"). Each data.frame has 2 columns:
\describe{
\item{pt_br}{The attribute in Brazilian Portuguese.}
\item{en}{The attribute in English.}
}
}
\usage{
data(map_translation)
}
\description{
A list of data.frames used by \code{faunabr::translate_faunabr()} function.
\code{faunabr::map_translation}.
}
\keyword{datasets}
