% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge Taxonomic Data from Catálogo Taxonômico da Fauna do Brasil}
\usage{
merge_data(
  path_data,
  version_data,
  solve_discrepancies = TRUE,
  translate = TRUE,
  encoding = "UTF-8",
  verbose = TRUE
)
}
\arguments{
\item{path_data}{(character) the base directory path where the downloaded
data is stored.}

\item{version_data}{(numeric) version of the Fauna do Brasil database
downloaded.}

\item{solve_discrepancies}{(logical) wheter to resolve inconsistencies
between species and subspecies information. When set to \code{TRUE} (default),
species information is updated based on unique data from subspecies. For
example, if a subspecies occurs in a certain state, it implies that the
species also occurs in that state.}

\item{translate}{(logical) whether to translate the original dataset
("lifeForm", "origin", "habitat", and "taxonRank") from Portuguese to
English. Default is \code{TRUE.}}

\item{encoding}{(character) The text string encoding to pass down to
\code{data.table::fread()}. Default is \code{"UTF-8"}.}

\item{verbose}{(logical) whether to display messages during function
execution. Set to TRUE to enable display, or FALSE to run silently. Default
is \code{TRUE}.}
}
\value{
The final dataset is built dynamically and saved directly to the matching
data directory as a compressed Gzip file named \code{"CompleteBrazilianFauna.gz"}.
}
\description{
Reads, processes, and merges the core components of the Catálogo Taxonômico
da Fauna do Brasil
}
\examples{
\dontrun{
merge_data(path_data = "data", version_data = "latest",
            solve_discrepancies = TRUE, translate = FALSE, verbose = TRUE)
}
}
