% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occurrences}
\alias{occurrences}
\title{Records of animal species}
\format{
A \code{data.frame} with 2798 rows and 3 variables:
\describe{
\item{species}{Species names (Panthera onca and Chaetomys subspinosus)}
\item{x}{Longitude}
\item{y}{Latitude}
}
}
\usage{
data(occurrences)
}
\description{
A dataset containing records of 2 plant species downloaded from GBIF. The
records were obtained with \code{plantR::rgbif2}
}
\references{
GBIF, 2024. florabr R package: Records of plant species. https://doi.org/10.15468/DD.QPGEB7
}
\keyword{datasets}
