% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_fauna}
\alias{world_fauna}
\title{SpatVector of the world countries}
\format{
A \code{SpatVector} with 258 geometries and 1 attribute:
\describe{
\item{name}{The name of the country (argentina, brazil, colombia, etc.)}
}
}
\usage{
world_fauna
}
\description{
A simplified and packed SpatVector of the world country polygons. The
spatial data was originally obtained from \code{rnaturalearth::ne_countries}.
Borders have been simplified by removing vertices of borders using
\code{terra::simplifyGeom}.It's necessary unpack the Spatvectos using
\code{terra::unwrap}

@usage data(world_fauna)
biomes <- terra::unwrap(world_fauna)
}
\keyword{datasets}
