% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fauna_version.R
\name{fauna_version}
\alias{fauna_version}
\title{Check if you have the latest version of Fauna do Brasil data available}
\usage{
fauna_version(data_dir)
}
\arguments{
\item{data_dir}{the directory where the data should be located.}
}
\value{
A message informing whether you have the latest version of Fauna do
Brasil available in the data_dir
}
\description{
This function checks if you have the latest version of the Fauna do Brasil
data available in a specified directory.
}
\examples{
#Check if there is a version of Fauna do Brasil data available in the
#current directory
fauna_version(data_dir = getwd())
}
